+++ api/src/main/java/org/ehrbase/api/service/EhrService.java
@@ -144,23 +144,6 @@ public interface EhrService extends BaseService {
     RevisionHistory getRevisionHistoryOfVersionedEhrStatus(UUID ehrUid);
 
     /**
-     * Reads the EHR entry from database and returns the ID of the root directory entry.
-     *
-     * @param ehrId - EHR id to find the directory for
-     * @return UUID of the root directory if existing
-     */
-    UUID getDirectoryId(UUID ehrId);
-
-    /**
-     * Removes the directory information from EHR table entry after deletion of the corresponding folder from
-     * folders table. If there were no such folder it will return a successful deletion.
-     *
-     * @param ehrId - Target EHR id
-     * @return Directory entry is now 'null'
-     */
-    boolean removeDirectory(UUID ehrId);
-
-    /**
      * Admin method to delete an EHR from the DB. See EHRbase Admin API specification for details.
      *
      * @param ehrId EHR to delete